+++ src/pages/cart/CartContent/CartContent.styled.ts
@@ -1,4 +1,4 @@
-import styled from '@emotion/styled';
+import styled from "@emotion/styled";
 
 export const Container = styled.div`
   width: 100%;
@@ -20,23 +20,6 @@ export const AllCheckBox = styled.div`
   margin-top: 36px;
   padding-bottom: 10px;
   box-shadow: 0 2px 4px -2px rgba(0, 0, 0, 0.1);
-`;
-
-export const OrderConfirmButton = styled.button`
-  width: 500px;
-  height: 64px;
-  position: fixed;
-  bottom: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  background-color: #000000;
-  color: #fff;
-  font-size: 16px;
-  font-weight: 700;
-
-  &:disabled {
-    background-color: #bebebe;
-  }
 `;
 
 export const ScrollContainer = styled.div`
