+++ src/pages/cart/CartContent/CartContent.tsx
@@ -4,8 +4,7 @@ import CartList from "./CartList/CartList";
 import CheckBox from "@/shared/components/CheckBox/CheckBox";
 import * as S from "./CartContent.styled";
 import { useOrderList } from "../hooks/useCartItem";
-import { useNavigate } from "react-router-dom";
-import { ROUTES } from "@/shared/config/routes";
+import OrderConfirmButton from "./CartList/OrderConfirmButton/OrderConfirmButton";
 
 export default function CartContent() {
   const {
@@ -20,7 +19,6 @@ export default function CartContent() {
     addSelectedItem,
     removeSelectedItem,
   } = useOrderList();
-  const navigate = useNavigate();
 
   if (isLoading) {
     return <div>로딩중</div>;
@@ -33,15 +31,6 @@ export default function CartContent() {
   if (!cartItems?.length) {
     return <EmptyCartContainer />;
   }
-
-  const handleOrderConfirmButtonClick = () => {
-    navigate(ROUTES.ORDER_SUCCESS, {
-      state: {
-        orderList,
-        orderTotalPrice,
-      },
-    });
-  };
 
   return (
     <S.Container>
@@ -64,13 +53,10 @@ export default function CartContent() {
         />
         <PriceContainer orderTotalPrice={orderTotalPrice} />
       </S.ScrollContainer>
-      <S.OrderConfirmButton
-        disabled={!orderList.length}
-        type="button"
-        onClick={handleOrderConfirmButtonClick}
-      >
-        주문 확인
-      </S.OrderConfirmButton>
+      <OrderConfirmButton
+        orderList={orderList}
+        orderTotalPrice={orderTotalPrice}
+      />
     </S.Container>
   );
 }
